requirements.txt
opencv-python
numpy
sounddevice
soundfile
torch
vosk
faster-whisper
transformers
langchain-ollama
langchain-community
pillow
